requirements.txt
msal==1.24.1
python-dotenv==1.0.0